+++ lib/rubocop/cop/doctolib_cops.rb
@@ -1,3 +1,4 @@
 # frozen_string_literal: true
 require_relative 'doctolib/no_asymmetrical_pundit_after_actions'
+require_relative 'doctolib/no_unsafe_background_migration'
 require_relative 'doctolib/one_operation_per_migration'